使用していない値の削除
