+++ kadai-4/package.json
@@ -8,7 +8,6 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.21.1",
     "core-js": "^3.6.5",
     "firebase": "^8.3.0",
     "vue": "^2.6.11",
