+++ kadai-4/src/config/firebase.config.js
@@ -1,0 +1,10 @@
+const firebaseConfig = {
+  apiKey: 'AIzaSyD-8X_eLWbZ-XW0tanR2RnUHi0hOtQPSrk',
+  authDomain: 'vue-kadai-4-6e35a.firebaseapp.com',
+  projectId: 'vue-kadai-4-6e35a',
+  storageBucket: 'vue-kadai-4-6e35a.appspot.com',
+  messagingSenderId: '921267101967',
+  appId: '1:921267101967:web:739b9992ed396f56665e7a',
+  measurementId: 'G-KG2MPX74TE'
+}
+export default firebaseConfig
